requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
distro==1.9.0
fonttools==4.53.0
gitdb==4.0.11
GitPython==3.1.43
groq==0.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.7
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.5
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.0
mdurl==0.1.2
more-itertools==10.3.0
narwhals==1.10.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
premailer==3.10.0
protobuf==5.28.2
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.9.3
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.39.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
watchdog==5.0.3
yagmail==0.15.293
